refactor: extract internal/provisioning package from internal/sensors
Moves provisioning store, postgres implementation, errors, and service
into internal/provisioning. Sensors package keeps thin aliases/wrappers
for backward compat.

Closes #100
